requirements.txt
aiogram==3.4.1
flask==2.3.3
python-dotenv==1.0.1
httpx==0.27.0
